Final Update - Step - Four - 22-02-2016

+++ src/in/jdsoft/studentmanagement/servlets/GroupServlet.java
@@ -2,12 +2,14 @@ package in.jdsoft.studentmanagement.servlets;
 
 import java.io.IOException;
 import java.util.ArrayList;
-
+import java.util.Map;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import com.google.gson.Gson;
 
 import in.jdsoft.studentmanagement.controller.GroupController;
 import in.jdsoft.studentmanagement.controller.GroupMappingController;
@@ -22,24 +24,42 @@ public class GroupServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		response.setContentType("text/html");
 		
-		Group group=new Group();
-		GroupController groupController=new GroupController();
-		GroupMappingController groupMappingController=new GroupMappingController();
-		ArrayList<CourseSemester> courseSemesters=new ArrayList<>();
-		
-		String groupName=request.getParameter("group-name");
-		String courseSemesterId[]=request.getParameterValues("courseSemList");
-		
-		for(String courseSemId:courseSemesterId){
-			CourseSemester courseSemester=new CourseSemester();
-			courseSemester.setCourseSemesterId(Integer.parseInt(courseSemId));
-		    courseSemesters.add(courseSemester);	
-		}
-		
-		if(request.getParameter("save")!=null){
-			group.setGroupName(groupName);
-			groupMappingController.addGroupMapping(groupController.addGroup(group), courseSemesters);
-		    response.sendRedirect("groups.jsp");
+		if(request.getParameter("action").equals("save")){
+			Group group=new Group();
+			GroupController groupController=new GroupController();
+			GroupMappingController groupMappingController=new GroupMappingController();
+			ArrayList<CourseSemester> courseSemesters=new ArrayList<>();
+			String groupName=request.getParameter("group-name");
+			String courseSemesterId[]=request.getParameterValues("courseSemList");
+			
+			
+			for(String courseSemId:courseSemesterId){
+				CourseSemester courseSemester=new CourseSemester();
+				courseSemester.setCourseSemesterId(Integer.parseInt(courseSemId));
+			    courseSemesters.add(courseSemester);	
+			}
+			      group.setGroupName(groupName);
+			      groupMappingController.addGroupMapping(groupController.addGroup(group), courseSemesters);
+			      groupController.addGroup(group);
+				 
+	    }
+		else if(request.getParameter("action").equals("update")){
+			Group group=new Group();
+			GroupController groupController=new GroupController();
+			GroupMappingController groupMappingController=new GroupMappingController();
+			ArrayList<CourseSemester> courseSemesters=new ArrayList<>();
+			String groupName=request.getParameter("editgroup-name");
+			String courseSemesterId[]=request.getParameterValues("courseSemList");
+			
+			
+			for(String courseSemId:courseSemesterId){
+				CourseSemester courseSemester=new CourseSemester();
+				courseSemester.setCourseSemesterId(Integer.parseInt(courseSemId));
+			    courseSemesters.add(courseSemester);	
+			}
+			      group.setGroupName(groupName);
+			      groupMappingController.addGroupMapping(groupController.addGroup(group), courseSemesters);
+			      groupController.updateGroup(group);
 		}
 		
 		}
@@ -47,12 +67,27 @@ public class GroupServlet extends HttpServlet {
 	  
 	   protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		//doPost(request, response);
-		int groupId=Integer.parseInt(request.getParameter("groupid"));
-		Group group=new Group();
-		GroupController groupController=new GroupController();
-		group.setGroupId(groupId);
-	    groupController.deleteGroup(group);
-		response.sendRedirect("groups.jsp");
+		   
+		   if(request.getParameter("action").equals("delete")){
+			   int groupId=Integer.parseInt(request.getParameter("groupid"));
+			   Group group=new Group();
+				GroupController groupController=new GroupController();
+				group.setGroupId(groupId);
+			    groupController.deleteGroup(group);
+				response.sendRedirect("groups.jsp");
+			}
+			else if(request.getParameter("action").equals("reterive")){
+				int groupId=Integer.parseInt(request.getParameter("groupid"));
+				GroupController groupController=new GroupController();
+				Map<String,Object> groupDetails= groupController.groupDetailsFromId(groupId);
+				String json = null;
+	            json = new Gson().toJson(groupDetails);
+	            
+	            response.setContentType("application/json");
+	            response.getWriter().write(json);
+				
+			}
+		
 		
 	   }
 }
